# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for HRD Attendance System",
        "title": "HRD Attendance API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/tenant-setting": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Tenant Setting"
                ],
                "summary": "Get Tenant Setting",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.TenantSetting"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenant Setting"
                ],
                "summary": "Update Tenant Setting",
                "parameters": [
                    {
                        "description": "Tenant Setting",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TenantSetting"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.TenantSetting"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/tenants": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenant"
                ],
                "summary": "Get All Tenants",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Tenant"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenant"
                ],
                "summary": "Create Tenant",
                "parameters": [
                    {
                        "description": "Tenant Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Tenant"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Tenant"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/tenants/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenant"
                ],
                "summary": "Get Tenant By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tenant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Tenant"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Tenant": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tenant_settings": {
                    "$ref": "#/definitions/model.TenantSetting"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "model.TenantSetting": {
            "type": "object",
            "properties": {
                "allow_multiple_check": {
                    "type": "boolean",
                    "example": false
                },
                "allow_remote": {
                    "type": "boolean",
                    "example": false
                },
                "clock_in_end_time": {
                    "type": "string",
                    "example": "09:00"
                },
                "clock_in_start_time": {
                    "type": "string",
                    "example": "07:00"
                },
                "clock_out_end_time": {
                    "type": "string",
                    "example": "23:00"
                },
                "clock_out_start_time": {
                    "type": "string",
                    "example": "16:00"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-04-07T13:21:24Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "late_after_minute": {
                    "type": "integer",
                    "example": 480
                },
                "max_radius_meter": {
                    "type": "number",
                    "example": 100
                },
                "office_latitude": {
                    "type": "number",
                    "example": -6.1339179
                },
                "office_longitude": {
                    "type": "number",
                    "example": 106.8329504
                },
                "require_location": {
                    "type": "boolean",
                    "example": true
                },
                "require_selfie": {
                    "type": "boolean",
                    "example": true
                },
                "tenant": {
                    "$ref": "#/definitions/model.Tenant"
                },
                "tenant_id": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-04-07T13:21:24Z"
                }
            }
        },
...
        "modelDto.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}